Refactor pod spec build
